feat: created welcome page with counter
